Cleaned up unused code
Removed: Unneeded files in World/Definitions folder and put the enumerator content in a Definition.cs file and the classes in their appropriate files.

UPDATED: Character.cs and made some fields public so they can be accessed.

Also added a Channel class with a list to hold channel information, just so you don't spam Join a channel you're already in..