Removed new line from the ReceiveChatMessage function on Info, Warning and Error messages.
Updated Character fields and added 'HasVerified'.

Added HandleLogonVerifyWorld opcode handle to ensure that the player has been verified.

ReceiveChatMessage will not start until the player has been Verified into the world.

Added a WorldMgr to the Client. This is mainly for simplification so we don't have to keep going through variables.