TicketForm now shows tickets that aren't closed
If you select a ticket in the list box, the information will go into the boxes below the list box.

Fixed the result not showing each ticket

Added EmuDevs information at the top of ChatHandler.cs in the Client since I forgot to do that.

Removed: TicketMgr from GMHelper to the Client

Edited what doesn't need to be selected from the gm_tickets table in PreparedResult file